LayoutTests/accessibility/mac/internal-link-when-document-has-fragment-expected.txt

Name
Name 2








Linked Element 1
Linked Element 2
This test that if a page's URL has a fragment appended to itself, it will still be able to find internal links within itself.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS body.url.indexOf('#name1') > -1 is true
PASS linkedElement.stringValue is 'AXValue: Linked Element 1'
PASS successfullyParsed is true

TEST COMPLETE
